% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComTreeOpt.r
\name{ComTreeOpt}
\alias{ComTreeOpt}
\title{Enhance user-specific community trees}
\usage{
ComTreeOpt(family, genus, species,
  megatree.uri = "https://www.dropbox.com/s/rc8jtmb6b2hdweo/R20160415.new?dl=0")
}
\arguments{
\item{family}{A list of families.}

\item{genus}{A list of genuses.}

\item{species}{A list of species. The list of families, genus and species must have the same number of entries.}

\item{megatree.uri}{URI to a megatree. The default value is the R20160415.new megatree (Gastauer & Meira-Neto, 2017).}
}
\value{
A phylogenetic tree containing the user-specified species, based on the Open Tree of Life.
}
\description{
Based on an interface with Open Tree of Life, the Community Tree Optimizer (ComTreeOpt) classifies families from user-specific lists into resolvable and non-resolvable families, builds resolved subtrees for resolvable families and inserts them in the megatree R20160415.new (Gastauer & Meira-Neto, 2017) previously pruned to species from non-resolvable families.
}
\examples{
data(FSN_species);
tree <- ComTreeOpt(family, genus, species);
plot(tree, show.node.label=T)
}
\references{
Gastauer, M. &  Meira-Neto, J. A. A. (2017) Updated angiosperm family tree for analyzing phylogenetic diversity and community structure. \emph{Acta Botanica Brasília}, vol.31, n.2, pp.191-198. DOI: \url{http://dx.doi.org/10.1590/0102-33062016abb0306}.
}
